requirements.txt
pytest~=8.1.1
allure-pytest~=2.13.5
selene>=2.0.0rc9
Appium-Python-Client~=4.0.0
python-dotenv~=1.0.1
requests~=2.31.0
pydantic_settings ~=2.2.1
pydantic~=2.6.4